feat: base page done

+++ app/Http/Controllers/EventController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreEventRequest;
 use App\Http\Requests\UpdateEventRequest;
 use App\Models\Event;
+use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HtmlSanitizer\HtmlSanitizer;
 
@@ -45,10 +46,36 @@ class EventController extends Controller
         if (isset($validated['description'])) {
             $validated['description'] = $sanitizer->sanitize($validated['description']);
         }
+        if (isset($validated['banner']) && $validated['banner'] instanceof UploadedFile) {
+            $banner = $validated['banner'];
+            unset($validated['banner']);
+        }
+
+        if (isset($validated['icon']) && $validated['icon'] instanceof UploadedFile) {
+            $icon = $validated['icon'];
+            unset($validated['icon']);
+        }
 
         $event = $request->user()->events()->create($validated);
 
-        return redirect()->route('events.show', $event);
+        if (isset($banner)) {
+            $fileName = $event->id.'.'.$banner->extension();
+            $banner->storeAs('banners', $fileName, 'public');
+            $validated['banner'] = $fileName;
+        }
+
+        if (isset($icon)) {
+            $fileName = $event->id.'.'.$icon->extension();
+            $icon->storeAs('icons', $fileName, 'public');
+            $validated['icon'] = $fileName;
+        }
+
+        if (isset($banner) || isset($icon)) {
+            $event->update($validated);
+        }
+
+
+        return redirect()->route('dashboard', $event)->with('success', 'Event created.');
     }
 
     /**
@@ -84,8 +111,22 @@ class EventController extends Controller
         if (isset($validated['description'])) {
             $validated['description'] = $sanitizer->sanitize($validated['description']);
         }
+
+        if (isset($validated['banner']) && $validated['banner'] instanceof UploadedFile) {
+            $fileName = $event->id.'.'.$validated['banner']->extension();
+            $validated['banner']->storeAs('banners', $fileName, 'public');
+            $validated['banner'] = $fileName;
+        }
+
+        if (isset($validated['icon']) && $validated['icon'] instanceof UploadedFile) {
+            $fileName = $event->id.'.'.$validated['icon']->extension();
+            $validated['icon']->storeAs('icons', $fileName, 'public');
+            $validated['icon'] = $fileName;
+        }
+
         $event->update($validated);
-        return redirect()->route('events.show', $event);
+
+        return redirect()->route('events.show', $event)->with('success', 'Event updated.');
     }
 
     /**
@@ -93,7 +134,22 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        if (Auth::user()->cannot('delete', $event)) {
+            return view('events.event', [
+                'event' => $event,
+            ])->with('error', 'You do not have permission to delete this event.');
+        }
+
+        if ($event->banner) {
+            unlink(storage_path('app/public/banners/'.$event->banner));
+        }
+
+        if ($event->icon) {
+            unlink(storage_path('app/public/icons/'.$event->icon));
+        }
+
         $event->delete();
-        return redirect()->route('events.index');
+
+        return redirect()->route('events.index')->with('success', 'Event deleted.');
     }
 }
